refactor(hardware): split tool detect and resolve
We have this ToolDetector thing. It listens for messages from the head
and returns the kinds of pipettes and gripper that are attached, polling
those devices for their information.

Let's refactor that to split those two things into a detect (which is
fast, and involves no further communication) and a resolve (which
involves lots more further communication). This lets us have a nice
async iterator for the detection and a lovely resolution mechanism that
can be called for each detect event, and then use all that stuff
precisely once for the oneshot.