refactor(hardware-control): add subsystem manager
If we want to hook a bunch of stuff off subsystems, we need a place to
handle them. We want to control the lifecycle of subsystems and their
updates, and that means we should get a lovely object to do this in.

This object has three jobs:
1. Manage the discovery of the "normal" subsystems on the network (i.e.
not stuff that's normally appearing and disappearing like tools), which
is what the ot3controller.probe_network() method used to do; this will
now use the opentrons_hardware.hardware_control.network
2. Manage those subsystems that do appear and disappear, and get the
additional information that they have. This uses the tools detector
infrastructure that we recently changed, triggered off the outcome of
network probes, and exposes and fuses all that information and turns it
into subsystem data.
3. Update those subsystems, including managing multiple ongoing updates
from multiple callers and handling the detection lifecycles of
subsystems as they get updated.

The primary interface here are the properties, because a key feature is
that this manages lifecycles _on its own_. You don't have to ask to
detect tools, because it's triggered off the head microcontroller
notifications; you don't have to re-refresh after an update, it will do
that.

There are some fun implementation details in here; for instance we use
the ever-exotic condition variable to allow for a bit of controlled
state that represents both whether we've completed a tool detection
cycle and whether the worker task got an error.

This commit doesn't stand alone because of the changes in the types
file; a followup will change how the ot3controller and hardware api work.