refactor(api): use subsystems as primary objects
Now that we have a subsystem manager, we can use it by turning the
hardware api into considering _subsystems_ as a new first-class object
alongside instruments.

subsystems are what get updated, not instruments or mounts; all those
subsystems being ready to run are what defines the api being ready to
operate, not several separate checks. There's not a lot of new
functionality here, mostly things get removed because they have
been added in new submodules.

Because we now rely on submodules as a first class object and subsystems
can be instruments, we can turn the instrument detection mechanism into
something that queries what subsystems are currently present - a flow
which does not require any canbus traffic and will always be correct
because tools are updated asynchronously from the head.