refactor(robot-server): add firmware updater
The firmware update manager will be the primary functionality driver of
the new subsystems endpoints. It handles the following needs:

- Clients should be able to see what needs updating
- Clients should be able to update things that need updating
- Clients should be able to make decisions based on updates that were
  started by other clients

The way to do this is by having a two-tier system of process management.
Processes can be either ongoing or completed. While processes are
ongoing, you can access them by the subsystem that they're updating.
Because hardware will raise exceptions if you try and start an update
process for a subsystem that's already being updated, there can only
ever be one ongoing update per subsystem.

When an update completes, whether it succeeds or fails, it is no longer
ongoing and goes over to be stored for the long term (until reboot) by
the id it was created with, and can later be retrieved by that ID.

The process flow for a client would be something like:
1. See that a subsystem needs updating
2. Begin an update
   2a. Either the update is in process, in which case we can get its
   status by subsystem, including its ID, or
   2b. The update begins and we can get its status and ID by subsystem
3. Watch the update by its ID, which will always be valid

Because 3 is by-id, you can't miss an event like the update ending while
you had your laptop closed or something. Because you can access an
update process initially by subsystem, you can display the status of an
update process that was started by someone else.

There's some fun implementation details here, like the use of handles to
make sure that callers can only interact with the ongoing process by one
end of a queue. The idea is anything outside this module either
interacts with a manager singleton or with process handles it gets from
the monitor singleton.