refactor(api,robot-server): Update fw in server
When we start the robot, we need to run any required firmware updates -
maybe we just got a system update, maybe somebody pushed something,
maybe the system is booting with a new instrument - and we want those
updates to be visible via HTTP. That means they need to be update
processes, and that means that they need to happen via the firmware
update manager in the server.

So let's hoist the initial update out of the hardware controller build
function and up into the robot server system startup tasks, which
requires refactoring them a bit to handle the new ot3-only behavior.

As a side-effect of this refactor, we should no longer be trying to
blink a nonexistent button on the OT-3, and we should have the server
become ready for interaction _much_ faster since it no longer waits for
firmware updates to be complete before sending its systemd ready notice.
While this might cause a weird time in the ODD interaction while
firmware updates are ongoing but the app seems fine, the solution to
that is the frontend changes to view and handle updates.