refactor(robot-server): get_ot3_hardware fixups
FastAPI actually evaluates the type hints on dependency functions at
runtime. This is a Really Big Problem for us because we can't rely on
OT3API being importable. It means that get_ot3_hardware can't be a
dependency; that also means it can't use dependencies and callers have
to pass in actual objects.

Also, you can't really do isinstance(some_thread_manager, OT3API)
because it's a thread manager not an OT3API; add a passthrough wrapper
for the thread manager that lets you ask what it's wrapping.