fix(api): disable motors during the pipette attach and detach position
We don't know what pipettes are attached on the hardware controller side until we actually detect
the pipette. This means that when the 96 channel gets attached for the first time the mount falls
because there isn't enough current on the head motors to support it. We should instead
auto-disengage all the ZA motors which will trigger the ebrake on the head board.