feat(robot-server): subsystems for bad instrs
The whole point of the subsystems interface is to have something to
present when an instrument is physically present but not actually able
to operate as an instrument for whatever reason. Since the need for the
subsystems is based on the need to handle instruments, and clients want
to be polling instruments, have /instruments return information about
the subsystem if an instrument needs an update or isn't ok so that
clients need to go look at the subsystems.