fix(shared-data): return heights for v3 pipettes
Update the return tip multipliers for the v3 pipettes to values from
hardware life testing... mostly.

This commit alters the v2 configs to have the proper values for all
tips, _but_ the protocol engine - which implements return tip - does not
use the v2 configs. It uses the v1 configs, which do not have separate
values for different tips.

For the pipettes that use multiple tips, we need to select just one
value. We want the distance from the bottom of each tip's ribs (and top)
to be the same, which means the multipliers are different. We select the
smallest multiplier; this makes larger tips too high; but that's better
than too low.

Since the numbers in the v2 configs are correct, whenever PE loads the
v2 configs everything will get fixed.