refactor(robot-server): split hw postinit
By putting postinit in a separate task, we can let the server come up
ready earlier. This lets us avoid issues with timeouts and allow clients
to see any details of the postinit tasks that we might want to expose -
for instance, the status of ongoing first-boot firmware updates for
connected subsystems.