hardware: handle USB comm errors
If we try and probe while the CANbus has an update going on, that's
fine, those devices just don't respond. But if we try and probe the USB
while the rear panel is updating, we can't find the usb device and will
get an exception. Since updates do a network probe when they complete,
if we have multiple concurrent updates then if the USB update doesn't
complete first the others will try and fail to probe it.

But really, all a communication error via USB means is that the probe
failed - that device is not present. So let's treat it like that.