refactor(hardware,api): probe_specific
When we probe the network, we send a broadcast ping and wait for
anything. We're askign the network in general what's present.

But with the subsystem manager, in a lot of situations what we really
want is to ask whether _specific devices_ are present, since we're
maintaining an estimate of what's present before then. We only want to
do the full broadcast when we start.

So add a new method called probe_specific, which instead of broadcasting
and returning _anything_ (and yelling at you if it misses what you
expect) will broadcast and look _only for responses from what you tell
it_, and not touch the rest of cached state.

This can then be used when a firmware update completes to prevent the
pings from looking for other devices in the middle of their own updates.