feat(app): Add move labware modal and implement into protocol flow (#12517)
* feat(app): Add move labware modal and implement into protocol flow

adds in the move labware intervention modal alongside the already existing pause intervention modal. The move labware modal will show up when a moveLabware command occurs and requires a manual move rather than using the gripper. Currently the modal will show the labware name, along with the movement being made (old -> new location) as well as the deck map.

The PR also wires up the modals to the protocol execution and should show the appropriate pause and move labware modals when those commands are run. The protocol can then be continued using the confirm button in the modal

closes RLAB-314
closes RLAB-315
closes RLAB-317