chore: do not build the app xplatform all the time (#12817)
It's good to run app builds to test if peoples' PRs work, but it's bad
if those app builds sporadically fail or take an hour even if they
succeed. Let's cut the number of times we do a full cross-platform app
build down by only doing the build if
- This is a tag
- This is edge
- This is release
- This is internal-release

In other cases, we'll only do the backend tests.