refactor: check devices more leniently after update
When we update, we check that all the devices that should have just
completed an update have come back up and are present on the network.
This lets us verify that the update succeeded. But they're rebooting
after an update, so we need to give them a little longer to start with a
nice retry system.

Also, add a mechanism to indicate that specific tools are no longer present.