refactor: exclusive can access for restarts
There's a period when our microcontroller applications and bootloaders
start where their can peripherals are running and accepting messages,
but no filters have been loaded. This usually isn't a problem, but it's
a really big problem when it's the bootloader, and at the same time
there's an update running and sending data for a different node - in
that case, the starting device gets a message for the other one and
reads it in. This doesn't cause corruption because we're checking
checksums and total packet counts, but it does make the update fail.

To fix that, we can introduce the concept of exclusive access to the can
messenger, since can messengers are shared. This pretty much boils down
to a lock. send() and ensure_send() take the lock all the time; but also
you can take the lock explicitly and prevent use of the canbus for some
longer time, and then use exclusive_send and exclusive_ensure_send to
avoid deadlocks.

By holding exclusive access in the updater while we're telling devices
to enter the bootloader, we should prevent any udpate data packets from
hitting the bus in the vulnerable period while a device is restarting.

There's some improvements that can be made here - there's no reason to
prevent concurrent access to the canbus when nothing's holding the
exclusive lock - but this works for now.