fix(robot-server): Better logging for postinit (#12895)
We have this awkward situation around logging, particularly for
exceptions, in hardware postinit. We really want to be able to print
stuff like where an exception occurred, but this is tragically really
annoying to do without being in the correct exception stack. There's a
way around this but it's weird: in the original exception stack, create
a special TracebackException instance, which will capture the right
things theoretically. However, that class _is not an exception_, and so
either everything has to return an exception or we have to wrap it in an
exception so we can raise it and use the old methods... so that's what
we do.