feat(shared-data): add error code data (#12905)
This json document gives basic details for defined error codes and
categories and some details. The details are for keeping people aware of
what the command means and to provide default information for a naive
client; they aren't full human-facing strings.

There are ts bindings for getting the details of an error given the code, 
and python bindings for structured exceptions.

This commit doesn't actually _use_ any of this; that will happen in followup PRs.