Update Thermocycler slots.
Describe which Flex slots it occupies, in addition to which OT-2 slots it occupies.

Describe special slot B1 for the Flex, in addition to special slot 7 for the OT-2.

Remove specific mention of DeckConflictError, which is not a public thing that users can import. Instead, just say, vaguely, that we'll raise "an error."