feat(robot-server): add error codes to responses
We can now express error codes in top level error responses and handle
and format them from the eneumerated errors that the rest of the system
will soon throw.

There's some interesting detail there around the MultiErrorResponse and
the ErrorResponse but with multiple errors captured inside of it... I'm
not sure what the right play is, but this is working out for me.

There are some response format changes for some of the messages; we can
fix these by making the exceptions that are being turned into the error
responses EnumeratedErrors that are properly formatted.