feat(hardware): Hardware error codes (#13009)
Now that we can express error codes in all our client interfaces, let's start raising them.

This PR makes the opentrons_hardware module use the new exceptions that have the new error codes. There's a couple new error codes, but mostly all this does is raise different exceptions in the same places.

The two exception is in the can messenger and the movegrouprunner.
CAN Messenger

The CAN messenger has this asyncio task that reads stuff asynchronously from the bus and passes it to registered listeners. In addition, if a message is an error message it would raise an exception. But, it's a task, so that exception doesn't go anywhere. #12963 made this exception just logged and swallowed but, like... why does it exist? So we removed it.
MoveGroupRunner

A similar issue happens in MoveGroupRunner, where a background task accumulates responses and then pops them on over to the caller. This works the same now, but we can also move the move stop condition checking inside that loop. We can also now properly handle having multiple errors in a move group!

Note: You need to run make setup-py again since now hardware depends on shared-data.