refactor(api): configure instrs only if needed
We've had this consistent problem with race conditions around
cache_instruments, which gets called when something hits /instruments.
It can reconfigure a bunch of internal data caches and also do things
like change the active currents on the axes. We used to do this pretty
often because it was the only time we actually checked what instruments
were connected. We configured everything unconditionally.

However, the app is basically always polling /instruments. That means
that if a process is running - a protocol, a calibration - some app is
going to hit /instruments and cause a current reconfiguration and
anything that requires a specific current will break, for instance. This
is bad!

The fix is to add in a concept from the OT2, of only running a system
reconfiguration if something actually changed since the last
cache_pipette. It's really easy to make this comparison, since we cache
stuff anyway - that's the point of cache pipette - and a lot of the
worker functions we call during the process actually return a bool to
let us know we can skip config anyway.