refactor(components): add higher and lower-level deck abstraction for Flex configuration
Create a RobotCoordinateSpace component that manages only scaling/reflection for the coordinate
system anchored in the bottom left. Create DeckSlotLocation component to rener a specific deck slot
location and bake path information into the component itself instead of deserialzing svgson. Create
a BaseDeck asbstraction that handles the basic rendering of a deck with labware, modules, and deck
labels