feat(hardware-testing): Added hardware testing usb package (#13140)
The hardware-testing package is used for testing various parts of the Flex, however in order to use it you have to have a computer that has all of our packages, and dependencies and has been set up to run our software. This poses some friction as it requires a lot to get up and running. This pr solves this problem by creating a hardware-testing package, that can be run from a USB drive plugged directly into the Flex.


- Added setup-usb-module to hardware-testing Makefile, which creates a USB package
- Added hardware-testing entry to scripts/python_build_utils.py so we can get the package version
- Added correct versioning based on git tags when producing sdist/wheel files