fix(app): poll instrument info for desktop protocol setup
To avoid scenarios where users complete a wizard flow launched from the protocol setup steps only to
return to a list item that still shows the old value, employ a background poll of the instruments
endpoint in addition to the on complete refetches.  this way updates are visible regardless of
whether ODD or Desktop was  used to complete the wizard.

Closes RQA-1139