fix(app): show robot network connection in app if able to communicate via that connection
adds a check to ensure a connected network status ip reported by a robot has a healthy response from
the app discovery-client. for example, robots connected to a different wifi network than the client
app should not show a wifi connection in the app, because the app cannot communicate with that robot
via wifi.

closes RQA-1150