refactor(app,app-shell,app-shell-odd): config v18
This removes the robot system update URLs from the config file. We need
to reshape them anyway, and we need them to be different depending on
the release channel, and the app updates aren't in there, so let's just
keep them in code in the node side

For app-shell-odd, this is pretty simple - we only ever download one
update set at a time since all we care about is flex; we determine which
update set by checking the interned project value.

For app-shell-odd, this is a bit complex because behavior will have to
change to download both updates. For now, to keep this change small,
we'll duplicate current behavior by downloading flex updates if this is
an internal-release and ot2 updates if this is production release.