refactor(app,app-shell-odd,app-shell): update actions
Changes the update file selection flow to use different files for
different robots.

Since we already download the two different kinds of file in the shell,
we need to figure out how to appropriately expose them to the app and
then how to appropriately select them. That means we need details
about the robot (which are available through redux on the app side) and
we need to send them to the app shell via one of the shell actions; and
we need to store the details about update availability separately for
the two robots (and be able to retrieve them separately).

First, storing and retrieving update details. This requires splitting
the redux state into two, one for ot2 and one for flex; each has the
same data about available system updates. The selectors that retrieve
that data now require robot names because we need to know what robot
kind to get data for.

Second, sending the right file. There are a couple possibilities for
where to do this, but there's one in particular that lets us bring
the user-specified-update-file flow (i.e. the one where you click a
button and pick a file from your filesytem) and the system automatic
flow closer together: specify the file via a read file action that
happens when you start an update.

This is a generalization of the READ_USER_FILE flow: the system file,
like the user file, would be stored in the redux mirror of the session;
a new action called READ_SYSTEM_FILE that carries the kind of robot gets
dispatched if you start a robot update without specifying your own file,
and the flows are the same from there (including the response FILE_INFO
action being the same between the two - it carries the same
information).

So the flow now looks like
- robotUpdate:START_ROBOT_UPDATE with EITHER a robot-type or a filepath
- if this was a filepath, robotUpdate:READ_USER_FILE to shell with path;
- if this was a robot-type, robotupdate:READ_SYSTEM_FILE to shell with
type;
- shell reads appropriate filekind and responds with
robotUpdate:USER_FILE_INFO
   - info is stored in redux session state
- robotUpdate:CREATE_SESSION
- (...the same flow...)
- robotUpdate:UPLOAD_FILE, which now always carries a filepath read from
the session state

And the migration workflow is unchanged.