chore: fix dev-mode updates
When we run the app in dev mode (i.e. with yarn electron
/path/to/app/dir), electron-updater still runs but takes its settings
from a dev-app-update.yml file instead of configuration interned by
electron-builder. It also reads the app current version directly from
package.json so you're always 0.0.0-dev.

By swapping out the app update yml file depending on which opentrons
project we're in we can point the dev app at the right place for its
updates depending on the value of the OPENTRONS_PROJECT variable. This
doesn't fix the current-version thing, but you can at least edit the
package.json file to fix that and now get the right update-prompting behavior.