refactor(app,app-shell,app-shell-odd): support updating both ot2 and flex (#13234)
The current app stack supports only one kind of update file and path. In internal-releases, we reuse that code for flex by pointing it at flex update media and flex robots... which just means the internal-release app can't update OT-2s with the automatically-distributed releases. It also renames from buildroot to robot update, fixes dev-mode updates, and lets you specify OPENTRONS_PROJECT in the ODD.

The new update flow works by:

- Downloading both ot-2 and flex update files, separately
- Keeping them around separately (in the app data dir)
- Keeping track of them in redux separately
  -  Including changing selectors to require specifying a robot name so we can get the right kind of update
- Pulling what kind of robot is getting updated from state when you start to update, and using the code paths previously only taken by user file robot updates (with the browse button) for everything

Note: This PR changes the download location for robot updates to be targeted on new infrastructure for flex launch. Specifically,

- internal-release robot software at ot3-development.builds.opentrons.com (sorry about the url, it's old). Flex builds can be seen here: https://ot3-development.builds.opentrons.com/ot3-oe/releases.json and OT-2 builds can be seen here: https://ot3-development.builds.opentrons.com/ot2-br/releases.json
- release robot software at builds.opentrons.com. This infra isn't up yet, but we can still merge this if we want

Also, I renamed all the redux and app-shell stuff from buildroot to robot-update.

Specifying OPENTRONS_PROJECT in the odd:

Add the ability to override OPENTRONS_PROJECT from the app-shell-odd
makefile and add a log to the app-shell-odd code for what url we're
lookinh for updates from and what we found.

Fixing dev-mode updates:

When we run the app in dev mode (i.e. with yarn electron
/path/to/app/dir), electron-updater still runs but takes its settings
from a dev-app-update.yml file instead of configuration interned by
electron-builder. It also reads the app current version directly from
package.json so you're always 0.0.0-dev.

By swapping out the app update yml file depending on which opentrons
project we're in we can point the dev app at the right place for its
updates depending on the value of the OPENTRONS_PROJECT variable. This
doesn't fix the current-version thing, but you can at least edit the
package.json file to fix that and now get the right update-prompting behavior.

More details about the update file Redux flow:

Changes the update file selection flow to use different files for
different robots.

Since we already download the two different kinds of file in the shell,
we need to figure out how to appropriately expose them to the app and
then how to appropriately select them. That means we need details
about the robot (which are available through redux on the app side) and
we need to send them to the app shell via one of the shell actions; and
we need to store the details about update availability separately for
the two robots (and be able to retrieve them separately).

First, storing and retrieving update details. This requires splitting
the redux state into two, one for ot2 and one for flex; each has the
same data about available system updates. The selectors that retrieve
that data now require robot names because we need to know what robot
kind to get data for.

Second, sending the right file. There are a couple possibilities for
where to do this, but there's one in particular that lets us bring
the user-specified-update-file flow (i.e. the one where you click a
button and pick a file from your filesytem) and the system automatic
flow closer together: specify the file via a read file action that
happens when you start an update.

This is a generalization of the READ_USER_FILE flow: the system file,
like the user file, would be stored in the redux mirror of the session;
a new action called READ_SYSTEM_FILE that carries the kind of robot gets
dispatched if you start a robot update without specifying your own file,
and the flows are the same from there (including the response FILE_INFO
action being the same between the two - it carries the same
information).

So the flow now looks like
- robotUpdate:START_ROBOT_UPDATE with EITHER a robot-type or a filepath
- if this was a filepath, robotUpdate:READ_USER_FILE to shell with path;
- if this was a robot-type, robotupdate:READ_SYSTEM_FILE to shell with
type;
- shell reads appropriate filekind and responds with
robotUpdate:USER_FILE_INFO
   - info is stored in redux session state
- robotUpdate:CREATE_SESSION
- (...the same flow...)
- robotUpdate:UPLOAD_FILE, which now always carries a filepath read from
the session state

And the migration workflow is unchanged.



Co-authored-by: Shlok Amin <shlokamin@me.com>