fix(api): normal home if fast home stalls
If a motor axis says it has a valid encoder position then when we home
we'll try and update the stepper from the encoder, fast move to near
home, and then home. But if the encoder position was wrong - if it
missed some overflows, for instance - then quite possibly that first
fast motion will fail, which right now would fail the home.

Luckily, the fast move is just an optimization. So let's catch a stall
error during the fast move and fall back to a normal home if it happens.
This should be very rare, which is good because "detecting a stall in
the fast move" really means "the fast move crashed into the side of the
robot", but you don't really notice since it's fast.

Closes RQA-1432