fix(api): base plunger speeds on working volume
When 62962c599b8b71bccc4c6b18a59a798f8b8989a3 added "liquid classes", we
had to change the way we look up various volumes, and we used the liquid
class max volume as the volume to look up ul/mm transforms for
determining speeds. The thing is, the liquid class max volume didn't
depend on the tip type, but the ul/mm transform did - so if you were
using a 50ul tip on a p1000, you'd try and get the ul/mm for 1000ul on a
50ul tip and get an error.

We can use the working volume instead, which is the maximum volume of
the tip, and fix this up but I think we need a better fix for this that
moves volume constraints into tip definitions.