fix: use max volume on OT-2
The pipette definitions for the OT-2 got converted to have different
information for tips, but those tips have ul/mm data all the way up to
the maximum volume of the pipette. That lets us stick with the max
volume.

That's also extremely important because the expected flow rates on OT-2
pipettes are those determined by looking up the flowrate from the
pipette nominal volume, so changing to using working volume would alter
the flowrates when using tips of different volumes than default. Thanks,
tests. Thests.